Db migration backup and restore done via npm package mysql-tools

+++ Courselect/server/migrations/20191207102208-courselect.js
@@ -11,18 +11,20 @@ var subject = require("../model/subject");
 var session = require("../model/session");
 var topic = require("../model/topic");
 var user = require("../model/user");
+var MysqlTools = require('mysql-tools');
+var config = require('../configs/config.json')
+var dbConfig = config.dbConfig;
 
 var mapping = require("../model/mapping");
 var async = require('async');
+
+// create database dump sql file
+var tool = new MysqlTools();
 
 /**
   * We receive the dbmigrate dependency from dbmigrate initially.
   * This enables us to not have to rely on NODE_PATH.
   */
-exports.check = function (db, callback) {
-  db.removeForeignKey(db, 'mapping', 'topic_fk', callback);
-};
-
 exports.setup = function (options, seedLink) {
   dbm = options.dbmigrate;
   type = dbm.dataType;
@@ -30,26 +32,48 @@ exports.setup = function (options, seedLink) {
 };
 
 exports.up = function (db, callback) {
-/*   db.createTable('topic', topic, callback);
-  db.createTable('session', session, callback);
-  db.createTable('subject', subject, callback);
-  db.createTable('course', course, callback);
-  db.createTable('mapping', mapping, callback); */
-  db.createTable('user', user, callback);
+  /*   db.createTable('topic', topic, callback);
+    db.createTable('session', session, callback);
+    db.createTable('subject', subject, callback);
+    db.createTable('course', course, callback);
+    db.createTable('mapping', mapping, callback);
+    db.createTable('user', user, callback);
+    db.createTable('userSession', userSession, callback); */
+  dbConfig.sqlFilePath = config.mysqlDumpPath
+
+  tool.restoreDatabase(dbConfig, function (error, output, message) {
+    if (error instanceof Error) {
+      console.log(error);
+    } else {
+      console.log(output);
+      console.log(message);
+    }
+  });
 };
-  
+
 exports.down = function (db, callback) {
-  db.removeForeignKey('mapping','topic_fk');
-  db.removeForeignKey('mapping','session_fk');
-  db.removeForeignKey('mapping','subject_fk');
-  db.removeForeignKey('mapping','course_fk');
-  db.dropTable('topic',callback);
-  db.dropTable('session',callback);
-  db.dropTable('subject',callback);
-  db.dropTable('course',callback);
-  db.dropTable('mapping',callback);
-  db.dropTable('user',callback);
-//  db.dropTable('migrations');
+  dbConfig.dumpPath = config.mysqlDumpPath
+  tool.dumpDatabase(dbConfig, function (error, output, message, dumpFileName) {
+    if (error instanceof Error) {
+      console.log(error);
+    } else {
+      console.log(output);
+      console.log(message);
+      console.log(dumpFileName);
+      db.removeForeignKey('mapping', 'topic_fk');
+      db.removeForeignKey('mapping', 'session_fk');
+      db.removeForeignKey('mapping', 'subject_fk');
+      db.removeForeignKey('mapping', 'course_fk');
+      db.dropTable('topic', callback);
+      db.dropTable('session', callback);
+      db.dropTable('subject', callback);
+      db.dropTable('course', callback);
+      db.dropTable('mapping', callback);
+      db.dropTable('user', callback);
+      return;
+    }
+  });
+  //mysqlDump({ connection: dbConfig, dumpToFile: mysqlDumpPath });
 };
 
 exports._meta = {
